Qunatum propagation seems okay!  We might have a functional spawning code!
Of course there are still issues: none of the quantum propagation info is currently printed out to output (not even amplitudes), because JSON can't handle complex numbers and I haven't gotten around to implementing HDF5 out put for the quantum bits yet.  This is the next thing to do.

After that...  Wigner initial conditions, interface with TeraChem, and extensive testing.